requirements.txt
### Im in Process of Redoing the requirements.txt file due to package conflicts
## Installed Again & Tested Working:
google-adk
python-dotenv
toolbox-core
requests
pandas
## Not Installed Yet / To Be Tested:
mcp
# pydantic
# selenium
# litellm
google-auth-oauthlib
firebase-admin
fastapi
uvicorn[standard]
# google-cloud-storage
# watchdog
jinja2
python-multipart
authlib
passlib[bcrypt]
pyjwt[crypto]
httpx